move project concerns to new class
